Add per OUI configuration which indicates whether or not to us SIP OPTIONS request as part of discovery process. This is to support SIP vendors which quietly ignore SIP OPTIONS requests. (XECS-2389)
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14993 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXcommons/src/main/java/org/sipfoundry/commons/discovery/DiscoveryAgentImpl.java
@@ -52,25 +52,33 @@ public class DiscoveryAgentImpl extends ActiveObject implements DiscoveryAgent {
     @Synchronous
     public void processPingResponse() {
     	if (discoveryService.isSIPVendor(targetAddress)) {
-            try {
-                ClientTransaction clientTransaction;
+    		if (discoveryService.supportsSIPOptions(targetAddress)) {
+    			try {
+    				ClientTransaction clientTransaction;
 
-                Request request = discoveryService.createOptionsRequest(targetAddress, 5060);
-                // Create the client transaction.
-                clientTransaction = sipProvider.getNewClientTransaction(request);
-                clientTransaction.setApplicationData(this);
-                // Send the request out.
-                clientTransaction.sendRequest();
-            } catch (TransactionUnavailableException e) {
-                e.printStackTrace();
-            } catch (SipException e) {
-                e.printStackTrace();
-            }
+    				Request request = discoveryService.createOptionsRequest(targetAddress, 5060);
+    				// Create the client transaction.
+    				clientTransaction = sipProvider.getNewClientTransaction(request);
+    				clientTransaction.setApplicationData(this);
+    				// Send the request out.
+    				clientTransaction.sendRequest();
+    			} catch (TransactionUnavailableException e) {
+    				e.printStackTrace();
+    			} catch (SipException e) {
+    				e.printStackTrace();
+    			}
+    		} else {
+    	        if (!terminated) {
+    	            discoveryService.addDiscovered(targetAddress, "");
+    	            ActiveObjectGroup activeObjectGroup = getActiveObjectGroup();
+    	            activeObjectGroup.deleteInstance(this);
+    	        }
+    		}
     	} else {
-            if (!terminated) {
-        	    ActiveObjectGroup activeObjectGroup = getActiveObjectGroup();
-        	    activeObjectGroup.deleteInstance(this);
-            }
+    		if (!terminated) {
+    			ActiveObjectGroup activeObjectGroup = getActiveObjectGroup();
+    			activeObjectGroup.deleteInstance(this);
+    		}
     	}
     }
